Cambios requeridos
